+++ src/test/java/com/github/havarunner/suite/suitedefinedinparent/AbstractSuiteMember.java
@@ -1,0 +1,7 @@
+package com.github.havarunner.suite.suitedefinedinparent;
+
+import com.github.havarunner.annotation.PartOf;
+
+@PartOf(Suite.class)
+class AbstractSuiteMember {
+}