+++ src/test/java/com/github/havarunner/suite/suitedefinedinparent/Suite.java
@@ -1,0 +1,12 @@
+package com.github.havarunner.suite.suitedefinedinparent;
+
+import com.github.havarunner.HavaRunnerSuite;
+
+class Suite implements HavaRunnerSuite<String> {
+    public String suiteObject() {
+        return "hello suite";
+    }
+
+    public void afterSuite() {
+    }
+}