+++ src/test/java/com/github/havarunner/suite/suitedefinedinparent/SuiteMember.java
@@ -1,0 +1,14 @@
+package com.github.havarunner.suite.suitedefinedinparent;
+
+import org.junit.Test;
+
+class SuiteMember extends AbstractSuiteMember {
+
+    SuiteMember(String suiteObject) {
+    }
+
+    @Test
+    public void hello() {
+        SuiteDefinedInParentTest.suiteMethodCalled = true;
+    }
+}